Move DKG components to separate `dkg.go` file
